feat(cli): apr cgp and apr qa-playbook, and 58 cgp tests that ran nothing
Two more crates whose whole command surface lived in a main.rs, reachable
only from their own binary. Both move to `<crate>::cli::{Cli, Commands,
run, dispatch}` with the binary as a shim, and apr's arm calls the SAME
dispatch. 26 more commands routed: 11 cgp + 15 qa.

  apr cgp doctor        -> probes the real machine: nvidia-smi 570.207,
                           CUDA 12.8, ncu, nsys, CUPTI, perf, renacer 0.10.2
  apr qa-playbook list  -> reads the real registry: "Total: 92 models"

Named `qa-playbook`, not `qa`: `apr qa` is already the falsifiable-gates
command that takes a model path. Two different tools, two names.

FOUND, and fixed here: crates/aprender-cgp/tests/{falsify,integration}.rs
built their subject as `cargo run -p cgp`. `cgp` is the [lib] name; the
PACKAGE is aprender-cgp. Every one of those 58 tests has died at cargo
package resolution since e54c1a3e1 (2026-06-12) -- the APR-MONO commit
that renamed the package and did not update its own test harness:

  $ cargo run -p cgp -- doctor ; echo rc=$?
  error: package(s) `cgp` not found in workspace
  rc=101

They were invisible because workspace-test runs
`cargo nextest run --workspace --lib` and neither file is on ci.yml's
beat list: dead AND unobserved.

Three of them reported PASS anyway, on cargo's own error text -- one
asserts only that output is non-empty, two carry
`if !output.status.success() { return; }`. Those are in #2496.

Reviving them immediately surfaced three real defects, all filed as #2496
rather than fixed here: the quant sweep prints a table whose every
measurement column is a dash; contract verification reports
"0 pass, 0 fail, 0 skip"; and a wall-clock assertion (623ms vs a 500ms
limit) of exactly the class #2425 removed three of. The harness fix is
safe to land now because these targets do not run in CI -- but they must
not be added to it until those three are fixed.

Also fixed, both mine from the previous commits in this branch:
  * a duplicated `_ => return None` arm in dispatch_analysis.rs that
    clippy flags as unreachable. `ci / lint` never reached it -- it had
    already failed on the doc lint above.
  * serde_json::json! in qa-cli expands to Result::unwrap, which this
    repo bans (GH-41). The diagnostic was real all along but charged to
    the bin target; moving the file into the lib is what surfaced it.

12 reach tests. Mutation: short-circuit dispatch_sibling_cli_commands to
None -> all 6 execution tests go RED, all 5 help-listing tests and the
sim parse-rejection test stay GREEN. That split is the point: the enum
stays fully wired and --help is byte-identical, so a reach test built
only on help output reports green while nothing executes.

Registry now 110 commands; FALSIFY-CLI-005 verified.

Refs #2495, #2496
