+++ Cargo.toml
@@ -257,6 +257,8 @@ aprender-db = { path = "crates/aprender-db", version = "0.63.0" }
 aprender-graph = { path = "crates/aprender-graph", version = "0.63.0" }
 aprender-rag = { path = "crates/aprender-rag", version = "0.63.0" }
 aprender-rag-cli = { path = "crates/aprender-rag-cli", version = "0.63.0" }
+aprender-qa-cli = { path = "crates/aprender-qa-cli", version = "0.63.0" }
+aprender-cgp = { path = "crates/aprender-cgp", version = "0.63.0" }
 aprender-data = { path = "crates/aprender-data", version = "0.63.0" }
 
 # Additional external deps needed by sub-crates
