+++ crates/apr-cli/Cargo.toml
@@ -238,6 +238,8 @@ aprender-rag-cli = { workspace = true }
 aprender-zram-cli = { workspace = true }
 # simular: hand-rolled parser, wired as a verbatim passthrough (apr sim ...)
 simular = { workspace = true }
+aprender-qa-cli = { workspace = true }
+aprender-cgp = { workspace = true }
 half = "2.4.1"
 # YAML config parsing (distillation configs, ALB-011)
 serde_yaml = { package = "serde_yaml_ng", version = "0.10" }
