+++ crates/apr-cli/src/commands_enum.rs
@@ -814,6 +814,16 @@ pub enum Commands {
     /// (was the `simular` binary)
     #[command(subcommand)]
     Sim(simular::cli::Command),
+
+    /// Compute-graph profiling: profile, bench, roofline, doctor
+    /// (was the `aprender-cgp` binary)
+    #[command(subcommand)]
+    Cgp(cgp::cli::Commands),
+
+    /// Model QA playbook runner: certify, run, score, parity
+    /// (was the `apr-qa` binary; `apr qa` is the falsifiable-gates command)
+    #[command(name = "qa-playbook", subcommand)]
+    QaPlaybook(aprender_qa_cli::cli::Commands),
 }
 
 /// Subcommands for `apr debug` (aprender#2377 finding 3).
