+++ crates/aprender-qa-cli/src/configuration.rs
@@ -13,253 +13,6 @@ fn setup_signal_handler() {
     }
 }
 
-/// Entry point that dispatches CLI subcommands to their handlers
-#[allow(clippy::too_many_lines)]
-fn main() {
-    setup_signal_handler();
-
-    let cli = Cli::parse();
-
-    match cli.command {
-        Commands::Certify {
-            all,
-            family,
-            tier,
-            kernel_class,
-            models,
-            output,
-            dry_run,
-            model_cache,
-            apr_binary,
-            auto_ticket,
-            ticket_repo,
-            no_integrity_check,
-            fail_fast,
-            oracle_enhance,
-        } => {
-            run_certification(
-                all,
-                family,
-                &tier,
-                kernel_class,
-                &models,
-                &output,
-                dry_run,
-                model_cache,
-                &apr_binary,
-                auto_ticket,
-                &ticket_repo,
-                no_integrity_check,
-                fail_fast,
-                oracle_enhance,
-            );
-        }
-        Commands::Run {
-            playbook,
-            output,
-            failure_policy,
-            fail_fast,
-            dry_run,
-            workers,
-            model_path,
-            timeout,
-            no_gpu,
-            skip_conversion_tests,
-            run_tool_tests,
-            profile_ci,
-            hf_parity,
-            hf_corpus_path,
-            hf_model_family,
-            no_integrity_check,
-            metadata_only,
-        } => {
-            // --fail-fast flag overrides --failure-policy
-            let effective_policy = if fail_fast {
-                "fail-fast".to_string()
-            } else {
-                failure_policy
-            };
-            run_playbook(
-                &playbook,
-                &output,
-                &effective_policy,
-                dry_run,
-                workers,
-                model_path,
-                timeout,
-                no_gpu,
-                skip_conversion_tests,
-                run_tool_tests,
-                profile_ci,
-                hf_parity,
-                &hf_corpus_path,
-                hf_model_family,
-                no_integrity_check,
-                metadata_only,
-            );
-        }
-        Commands::Tools {
-            model_path,
-            no_gpu,
-            output,
-            include_serve,
-        } => {
-            run_tool_tests(&model_path, no_gpu, &output, include_serve);
-        }
-        Commands::Generate {
-            model,
-            count,
-            format,
-        } => {
-            generate_scenarios(&model, count, &format);
-        }
-        Commands::Score { evidence, model } => {
-            calculate_score(&evidence, &model);
-        }
-        Commands::Report {
-            evidence,
-            output,
-            formats,
-            model,
-        } => {
-            generate_report(&evidence, &output, &formats, &model);
-        }
-        Commands::List { size } => {
-            list_models(size.as_deref());
-        }
-        Commands::LockPlaybooks { dir, output } => match generate_lock_file(&dir, &output) {
-            Ok(0) => {
-                eprintln!("Error: No playbook files found in {}", dir.display());
-                std::process::exit(1);
-            }
-            Ok(count) => println!("Locked {count} playbook(s) → {}", output.display()),
-            Err(e) => {
-                eprintln!("Error generating lock file: {e}");
-                std::process::exit(1);
-            }
-        },
-        Commands::Tickets {
-            evidence,
-            repo,
-            black_swans_only,
-            min_occurrences,
-            ticket_mode,
-        } => {
-            generate_tickets(
-                &evidence,
-                &repo,
-                black_swans_only,
-                min_occurrences,
-                &ticket_mode,
-            );
-        }
-        Commands::Parity {
-            model_family,
-            corpus_path,
-            logits_file,
-            prompt,
-            tolerance,
-            list,
-            self_check,
-        } => {
-            run_parity_check(
-                &model_family,
-                &corpus_path,
-                logits_file.as_deref(),
-                prompt.as_deref(),
-                &tolerance,
-                list,
-                self_check,
-            );
-        }
-        Commands::ExportCsv {
-            evidence_dir,
-            output,
-            append,
-        } => {
-            export_csv(&evidence_dir, &output, append);
-        }
-        Commands::ExportEvidence {
-            source,
-            output_dir,
-            model,
-            family,
-            size,
-            playbook_name,
-            tier,
-        } => {
-            export_evidence(
-                &source,
-                &output_dir,
-                &model,
-                &family,
-                &size,
-                &playbook_name,
-                &tier,
-            );
-        }
-        Commands::Bootstrap {
-            family,
-            size,
-            hf_repo,
-            tier,
-            output,
-            contracts_path,
-            dry_run,
-        } => {
-            run_bootstrap(
-                &family,
-                &size,
-                &hf_repo,
-                &tier,
-                output.as_deref(),
-                &contracts_path,
-                dry_run,
-            );
-        }
-        Commands::ValidateContract {
-            model_path,
-            contract_path,
-            format,
-            critical_only,
-        } => {
-            validate_contract_command(
-                &model_path,
-                contract_path.as_deref(),
-                &format,
-                critical_only,
-            );
-        }
-        Commands::KernelCoverage {
-            architecture,
-            all,
-            models,
-            verify,
-            file_tickets,
-            output_dir,
-            trueno_path,
-            realizar_path,
-            format,
-            contracts_path,
-            bindings_path,
-        } => {
-            kernel_coverage_command(
-                architecture.as_deref(),
-                all,
-                models,
-                verify,
-                file_tickets,
-                &output_dir,
-                &trueno_path,
-                &realizar_path,
-                &format,
-                &contracts_path,
-                &bindings_path,
-            );
-        }
-    }
-}
-
 /// Print playbook execution status including model, workers, and test configuration
 #[allow(clippy::fn_params_excessive_bools)]
 #[allow(clippy::too_many_lines)]
